refactor: extract runtime reference helpers
